Rename mod reader to arrow
