+++ rust/parquet/src/lib.rs
@@ -30,10 +30,10 @@ pub use self::util::memory;
 
 #[macro_use]
 mod util;
+pub mod arrow;
 pub mod column;
 pub mod compression;
 mod encodings;
 pub mod file;
-pub mod reader;
 pub mod record;
 pub mod schema;
